requirements.txt
langchain
langchain-groq
python-dotenv
langchain-community
discord
requests
pillow